+++ src/contexts/FeatureFlagContext.ts
@@ -2,6 +2,7 @@ import { createContext } from 'react';
 import { FeatureToggles, urlUtil } from '@grafana/data';
 import { config } from '@grafana/runtime';
 import { FeatureName } from 'types';
+import { isArray } from 'lodash';
 
 interface FeatureFlagContextValue {
   featureToggles: FeatureToggles;
@@ -9,9 +10,12 @@ interface FeatureFlagContextValue {
 }
 
 export function isFeatureEnabled(name: FeatureName) {
-  const isEnabledThroughQueryParam = urlUtil.getUrlSearchParams()['features']?.includes(name);
-  // @ts-ignore
-  // the type definitions in grafana core for feature toggles aren't quite right
+  const featuresParam = urlUtil.getUrlSearchParams()['features'];
+  let isEnabledThroughQueryParam = false;
+  if (isArray(featuresParam)) {
+    const stringParams = featuresParam as string[];
+    isEnabledThroughQueryParam = stringParams.includes(name);
+  }
   return Boolean(config.featureToggles[name]) || isEnabledThroughQueryParam;
 }