+++ src/page/WelcomePage.tsx
@@ -21,6 +21,7 @@ import circledLoki from 'img/circled-loki.svg';
 import { CloudDatasourceJsonData } from 'datasource/types';
 import { isNumber } from 'lodash';
 import { OrgRole } from 'types';
+import { trackEvent, trackException } from 'analytics';
 
 const getStyles = (theme: GrafanaTheme) => {
   const textColor = theme.isDark ? colors.darkText : colors.lightText;
@@ -33,6 +34,7 @@ const getStyles = (theme: GrafanaTheme) => {
       left: 0;
       background: ${theme.isDark ? colors.darkThemeBlueBackground : colors.lightThemeBlueBackground};
       color: ${textColor};
+      z-index: 1;
     `,
     bannerBackground: css`
       background-image: url(${theme.isDark ? darkCircledSM : lightCircledSM});
@@ -128,8 +130,10 @@ export const WelcomePage: FC<Props> = () => {
   const logsDatasource = config.datasources[logsName] as DataSourceInstanceSettings<CloudDatasourceJsonData>;
   const stackId = meta?.jsonData?.stackId;
   const onClick = async () => {
+    trackEvent('provisionedSetupSubmit');
     if (!meta?.jsonData) {
       setError('Invalid plugin configuration');
+      trackException('provisionedSetupSubmitError: Invalid plugin configuration');
       return;
     }
     setLoading(true);
@@ -159,6 +163,7 @@ export const WelcomePage: FC<Props> = () => {
     } catch (e) {
       setError(e.data?.msg ?? e.data?.err);
       setLoading(false);
+      trackException(`provisionedSetupSubmitError: ${e.data?.msg ?? e.data?.err}`);
     }
   };