+++ src/types.ts
@@ -268,7 +268,7 @@ export interface CheckFormValues extends Omit<Check, 'settings' | 'labels' | 'al
   settings: SettingsFormValues;
   labels?: Label[];
   alertSensitivity: SelectableValue<AlertSensitivity>;
-  useFullMetrics: boolean;
+  publishAdvancedMetrics: boolean;
 }
 
 export interface Check extends BaseObject {
@@ -468,6 +468,7 @@ export enum CheckEnabledStatus {
 export enum CheckListViewType {
   Card,
   List,
+  Viz,
 }
 
 export enum HTTPCompressionAlgo {